Add remark feature
No way to add a remark to a profile.

Being able to add a remark to a profile will improve
user experience as they will be able to keep track of
more information.

Let's add a remark field to the Person class, and add
the functionality needed for a remark command.

New profiles created will have an empty remark, and
can only be edited using the remark command.